+++ Flask/src/Models/models.py
@@ -1,0 +1,14 @@
+from flask import Flask, render_template, url_for, request, redirect
+from flask_sqlalchemy import SQLAlchemy
+from datetime import datetime
+db = SQLAlchemy()
+
+
+class Todo(db.Model):
+    id = db.Column(db.Integer, primary_key = True)
+    content = db.Column(db.String(200), nullable = False)
+    completed = db.Column(db.Integer, default = 0)
+    date_created = db.Column(db.DateTime, default = datetime.utcnow)
+
+    def __repr__(self):
+        return '<Task %r>' % self.id
